requirements.txt
python-telegram-bot
daemons
requests
docopt
